bump version number, document changes.

+++ libxfs/xfs_dir_leaf.c
@@ -62,8 +62,8 @@ xfs_dir_ino_validate(xfs_mount_t *mp, xfs_ino_t ino)
 		XFS_AGINO_TO_INO(mp, agno, agino) == ino;
 	if (XFS_TEST_ERROR(!ino_ok, mp, XFS_ERRTAG_DIR_INO_VALIDATE,
 			XFS_RANDOM_DIR_INO_VALIDATE)) {
-		xfs_fs_cmn_err(CE_WARN, mp,
-			"Invalid inode number 0x%Lx\n", ino);
+		xfs_fs_cmn_err(CE_WARN, mp, "Invalid inode number 0x%Lx\n",
+				(unsigned long long) ino);
 		return XFS_ERROR(EFSCORRUPTED);
 	}
 	return 0;
